requirements.txt
streamlit
numpy
opencv-python-headless
pytesseract
Pillow
requests